requirements.txt
python-telegram-bot==20.7
matplotlib==3.8.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
